+++ waton/client/__init__.py
@@ -1,5 +1,6 @@
 """Client package public exports."""
 
+from .business import BusinessAPI
 from .client import WAClient
 
-__all__ = ["WAClient"]
+__all__ = ["WAClient", "BusinessAPI"]